+++ src/integration-test/groovy/com/causecode/content/blog/BlogControllerSpec.groovy
@@ -6,7 +6,6 @@
  * without modification, are not permitted.
  *//*
 
-
 package com.causecode.content.blog
 
 import spock.lang.*